+++ resources/views/proveedores/index.blade.php
@@ -7,9 +7,11 @@
             <input class="boton boton-principal" type="submit" value="Buscar" id="buscar">
         </form>
     
-        <div class="opciones">
-            <a href="/{{ strtolower($titulo).'/registrar' }}" class="boton boton-registrar">Registrar nuevo</a>
-        </div>
+        @if (Auth::user()->is_admin == '1')
+            <div class="opciones">
+                <a href="/{{ strtolower($titulo).'/registrar' }}" class="boton boton-registrar">Registrar nuevo</a>
+            </div>
+        @endif
     </div>
 
     <div class="table-container">
@@ -31,8 +33,10 @@
                         <td>
                             <div class="acciones">
                                 <a title="Consultar" class="fa-solid fa-eye" href="/{{ strtolower($titulo) }}/consultar?id={{ $dato->id }}"></a>
-                                <a title="Actualizar" class="fa-solid fa-pen-to-square" href="/{{ strtolower($titulo) }}/actualizar?id={{ $dato->id }}"></a>
-                                <a title="Eliminar" class="fa-solid fa-trash-can" onclick="return confirm('¿Seguro que desea eliminar el registro?')" href="/{{ strtolower($titulo) }}/eliminar?id={{ $dato->id }}"></a>
+                                @if (Auth::user()->is_admin == '1')
+                                    <a title="Actualizar" class="fa-solid fa-pen-to-square" href="/{{ strtolower($titulo) }}/actualizar?id={{ $dato->id }}"></a>
+                                    <a title="Eliminar" class="fa-solid fa-trash-can" onclick="return confirm('¿Seguro que desea eliminar el registro?')" href="/{{ strtolower($titulo) }}/eliminar?id={{ $dato->id }}"></a>
+                                @endif
                             </div>
                         </td>
                     </tr>
